+++ doc/changelog.rst
@@ -12,6 +12,7 @@ Unreleased
 Core
 ~~~~
 
+- Add `get_customer_name` for `Order`
 - Exclude images from product `get_public_media`
 - Add parameter to `PriceDisplayFilter` to specify tax display mode
 - Add soft deletion of categories
@@ -40,6 +41,7 @@ Localization
 Admin
 ~~~~~
 
+- Fix bug with empty customer names in order list view
 - Add warning when editing order with no customer contact
 - Show account manager info on order detail page
 - Remove "Purchased" checkbox from product images section